changed layout for our one
