+++ ALARM/app/src/main/java/com/example/alarm/TabFragment_Alarm.java
@@ -1,130 +1,86 @@
 package com.example.alarm;
 
-import android.app.AlarmManager;
-import android.app.DatePickerDialog;
-import android.app.PendingIntent;
-import android.content.Intent;
+import android.animation.ArgbEvaluator;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.DatePicker;
-import android.widget.TimePicker;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.viewpager.widget.ViewPager;
 
 import com.example.alarm.alarm_card.Adapter;
 import com.example.alarm.alarm_card.Model;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.List;
-import java.util.Locale;
 
 public class TabFragment_Alarm extends Fragment {
-    private View view;
-    List<Model> models;
-    Adapter adapter;
-    // 알람 시간
-    private Calendar calendar;
+
     static int cnt=3;
-    private TimePicker timePicker;
-    public static TabFragment_Alarm newinstance()
-    {
-        TabFragment_Alarm tabFragmentAlarm =new TabFragment_Alarm();
-        return tabFragmentAlarm;
-    }
+    ViewPager alarm_viewpager;
+    Adapter adapter;
+    List<Model> models;
+    Integer[] colors = null;
+    ArgbEvaluator argbEvaluator = new ArgbEvaluator();
+
     @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedinstanceState){
-        view=inflater.inflate(R.layout.tab_fragment_alarm,container,false);
-        this.calendar = Calendar.getInstance();
-        // 현재 날짜 표시
-        displayDate();
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.tab_fragment_info2, null);
 
-        this.timePicker = view.findViewById(R.id.timePicker);
-        //Calender,알람버튼에 리스너 추가
-        view.findViewById(R.id.btnCalendar).setOnClickListener(mClickListener);
-        view.findViewById(R.id.btnAlarm).setOnClickListener(mClickListener);
+        if(getArguments()!=null)
+        {
+            cnt=getArguments().getInt("cnt");
+            Toast.makeText(view.getContext(), "Alarm : " , Toast.LENGTH_LONG).show();
+        }
+        Toast.makeText(view.getContext(), "Alarm : " , Toast.LENGTH_LONG).show();
+        models = new ArrayList<>();
+        for(int i = 1; i< TabFragment_Alarm.cnt + 1; i++)
+        {
+            models.add(new Model(R.drawable.brochure, Integer.toString(i), Integer.toString(i)+"번째알람"));
+        }
+
+        adapter = new Adapter(models, getContext());
+
+        alarm_viewpager=view.findViewById(R.id.alarm_viewpager);
+        alarm_viewpager.setAdapter(adapter);
+        alarm_viewpager.setPadding(130, 0, 130, 0);
+
+        Integer[] colors_temp = {
+                getResources().getColor(R.color.color1),
+                getResources().getColor(R.color.color2),
+                getResources().getColor(R.color.color3),
+                getResources().getColor(R.color.color4)
+        };
+
+        colors = colors_temp;
+
+        alarm_viewpager.setOnPageChangeListener(new ViewPager.OnPageChangeListener() {
+            @Override
+            public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
+                if(position<(adapter.getCount() - 1) && position< (colors.length - 1)){
+                    alarm_viewpager.setBackgroundColor((Integer) argbEvaluator.evaluate(positionOffset, colors[position], colors[position+1]));
+                }else{
+                    alarm_viewpager.setBackgroundColor(colors[colors.length - 1]);
+                }
+            }
+
+            @Override
+            public void onPageSelected(int position) {
+
+            }
+
+            @Override
+            public void onPageScrollStateChanged(int state) {
+
+            }
+        });
+
+
         return view;
     }
-
-
-    /* 날짜 표시 */
-    private void displayDate() {
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd", Locale.getDefault());
-        //((TextView) findViewById(R.id.txtDate)).setText(format.format(this.calendar.getTime()));
-    }
-
-    /* DatePickerDialog 호출 */
-    private void showDatePicker() {
-        DatePickerDialog dialog = new DatePickerDialog(view.getContext(), new DatePickerDialog.OnDateSetListener() {
-            @Override
-            public void onDateSet(DatePicker view, int year, int month, int dayOfMonth) {
-                // 알람 날짜 설정
-                calendar.set(Calendar.YEAR, year);
-                calendar.set(Calendar.MONTH, month);
-                calendar.set(Calendar.DATE, dayOfMonth);
-
-                // 날짜 표시
-                displayDate();
-            }
-        }, this.calendar.get(Calendar.YEAR), this.calendar.get(Calendar.MONTH), this.calendar.get(Calendar.DAY_OF_MONTH));
-
-        dialog.show();
-    }
-
-    /* 알람 등록 */
-    private void setAlarm() {
-        // 알람 시간 설정
-        this.calendar.set(Calendar.HOUR_OF_DAY, this.timePicker.getHour());
-        this.calendar.set(Calendar.MINUTE, this.timePicker.getMinute());
-        this.calendar.set(Calendar.SECOND, 0);
-
-        // 현재일보다 이전이면 등록 실패
-        if (this.calendar.before(Calendar.getInstance())) {
-            Toast.makeText(view.getContext(), "알람시간이 현재시간보다 이전일 수 없습니다.", Toast.LENGTH_LONG).show();
-            return;
-        }
-
-        // Receiver 설정
-        Intent intent = new Intent(view.getContext(), AlarmReceiver.class);
-        Fragment fragment =new TabFragment_Info();
-        Bundle bundle =new Bundle();
-        bundle.putInt("cnt",cnt);
-        fragment.setArguments(bundle);
-
-        PendingIntent pendingIntent = PendingIntent.getBroadcast(view.getContext(), cnt++, intent, PendingIntent.FLAG_UPDATE_CURRENT);
-
-        // 알람 설정
-        AlarmManager alarmManager = (AlarmManager) getActivity().getSystemService(view.getContext().ALARM_SERVICE);
-        alarmManager.set(AlarmManager.RTC_WAKEUP, this.calendar.getTimeInMillis(), pendingIntent);
-
-        // Toast 보여주기 (알람 시간 표시)
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.getDefault());
-        Toast.makeText(view.getContext(), "Alarm : "+Integer.toString(cnt) + format.format(calendar.getTime()), Toast.LENGTH_LONG).show();
-    }
-
-
-    View.OnClickListener mClickListener = new View.OnClickListener() {
-        @Override
-        public void onClick(View v) {
-            switch (v.getId()) {
-                case R.id.btnCalendar:
-                    // 달력
-                    showDatePicker();
-
-                    break;
-                case R.id.btnAlarm:
-                    // 알람 등록
-                    setAlarm();
-
-                    break;
-            }
-        }
-    };
 }
