+++ ALARM/app/src/main/res/layout/activity_main.xml
@@ -29,7 +29,7 @@
         <com.google.android.material.tabs.TabItem
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="setting" />
+            android:text="Stats" />
 
 
     </com.google.android.material.tabs.TabLayout>
