+++ ALARM/app/src/main/res/layout/tab_fragment_info2.xml
