+++ ALARM/app/src/main/res/layout/tab_fragment_statistic.xml
@@ -8,5 +8,5 @@
         android:id="@+id/textView3"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="fragment3" />
+        android:text="fragment_stat" />
 </LinearLayout>
